Se realizó una limpieza y optimización del backend para mejorar la estabilidad del sistema. Los cambios principales incluyen:

Persistencia: Corrección de mapeos JPA en entidades y optimización de consultas en Repositorios (búsquedas por DPI).

Lógica de Negocio: Refactorización de la capa de Servicios para asegurar el manejo correcto de stock y estados de registros.

API REST: Actualización de controladores para estandarizar las respuestas HTTP y la comunicación con la capa de servicio.

+++ src/main/java/com/javiercajchun/kinalapp/repository/UsuarioRepository.java
@@ -2,7 +2,8 @@ package com.javiercajchun.kinalapp.repository;
 
 import com.javiercajchun.kinalapp.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
 }
